Move dataset API to services pkg
